+++ tests/test_prediction.py
@@ -7,9 +7,11 @@ from akkudoktoreos.prediction.elecpriceenergycharts import ElecPriceEnergyCharts
 from akkudoktoreos.prediction.elecpricefixed import ElecPriceFixed
 from akkudoktoreos.prediction.elecpriceimport import ElecPriceImport
 from akkudoktoreos.prediction.elecpricetibber import ElecPriceTibber
+from akkudoktoreos.prediction.feedintariffakkudoktor import FeedInTariffAkkudoktor
 from akkudoktoreos.prediction.feedintariffenergycharts import FeedInTariffEnergyCharts
 from akkudoktoreos.prediction.feedintarifffixed import FeedInTariffFixed
 from akkudoktoreos.prediction.feedintariffimport import FeedInTariffImport
+from akkudoktoreos.prediction.feedintarifftibber import FeedInTariffTibber
 from akkudoktoreos.prediction.loadakkudoktor import (
     LoadAkkudoktor,
     LoadAkkudoktorAdjusted,
@@ -48,8 +50,10 @@ def forecast_providers():
         ElecPriceFixed(),
         ElecPriceImport(),
         FeedInTariffEnergyCharts(),
+        FeedInTariffAkkudoktor(),
         FeedInTariffFixed(),
         FeedInTariffImport(),
+        FeedInTariffTibber(),
         LoadAkkudoktor(),
         LoadAkkudoktorAdjusted(),
         LoadVrm(),
@@ -102,22 +106,24 @@ def test_provider_sequence(prediction):
     assert isinstance(prediction.providers[3], ElecPriceFixed)
     assert isinstance(prediction.providers[4], ElecPriceImport)
     assert isinstance(prediction.providers[5], FeedInTariffEnergyCharts)
-    assert isinstance(prediction.providers[6], FeedInTariffFixed)
-    assert isinstance(prediction.providers[7], FeedInTariffImport)
-    assert isinstance(prediction.providers[8], LoadAkkudoktor)
-    assert isinstance(prediction.providers[9], LoadAkkudoktorAdjusted)
-    assert isinstance(prediction.providers[10], LoadVrm)
-    assert isinstance(prediction.providers[11], LoadImport)
-    assert isinstance(prediction.providers[12], PVForecastAkkudoktor)
-    assert isinstance(prediction.providers[13], PVForecastVrm)
-    assert isinstance(prediction.providers[14], PVForecastPVNode)
-    assert isinstance(prediction.providers[15], PVForecastForecastSolar)
-    assert isinstance(prediction.providers[16], PVForecastSolcast)
-    assert isinstance(prediction.providers[17], PVForecastImport)
-    assert isinstance(prediction.providers[18], WeatherBrightSky)
-    assert isinstance(prediction.providers[19], WeatherClearOutside)
-    assert isinstance(prediction.providers[20], WeatherOpenMeteo)
-    assert isinstance(prediction.providers[21], WeatherImport)
+    assert isinstance(prediction.providers[6], FeedInTariffAkkudoktor)
+    assert isinstance(prediction.providers[7], FeedInTariffFixed)
+    assert isinstance(prediction.providers[8], FeedInTariffImport)
+    assert isinstance(prediction.providers[9], FeedInTariffTibber)
+    assert isinstance(prediction.providers[10], LoadAkkudoktor)
+    assert isinstance(prediction.providers[11], LoadAkkudoktorAdjusted)
+    assert isinstance(prediction.providers[12], LoadVrm)
+    assert isinstance(prediction.providers[13], LoadImport)
+    assert isinstance(prediction.providers[14], PVForecastAkkudoktor)
+    assert isinstance(prediction.providers[15], PVForecastVrm)
+    assert isinstance(prediction.providers[16], PVForecastPVNode)
+    assert isinstance(prediction.providers[17], PVForecastForecastSolar)
+    assert isinstance(prediction.providers[18], PVForecastSolcast)
+    assert isinstance(prediction.providers[19], PVForecastImport)
+    assert isinstance(prediction.providers[20], WeatherBrightSky)
+    assert isinstance(prediction.providers[21], WeatherClearOutside)
+    assert isinstance(prediction.providers[22], WeatherOpenMeteo)
+    assert isinstance(prediction.providers[23], WeatherImport)
 
 
 def test_provider_by_id(prediction, forecast_providers):
@@ -139,7 +145,9 @@ def test_prediction_repr(prediction):
     assert "ElecPriceFixed" in result
     assert "ElecPriceImport" in result
     assert "FeedInTariffFixed" in result
+    assert "FeedInTariffAkkudoktor" in result
     assert "FeedInTariffImport" in result
+    assert "FeedInTariffTibber" in result
     assert "LoadAkkudoktor" in result
     assert "LoadVrm" in result
     assert "LoadImport" in result